+++ js/inputmask.js
@@ -69,6 +69,7 @@
 			nojumpsThreshold: 0, //start nojumps as of
 			keepStatic: undefined, //try to keep the mask static while typing. Decisions to alter the mask will be posponed if possible - undefined see auto selection for multi masks
 			positionCaretOnTab: false, //when enabled the caret position is set after the latest valid position on TAB
+			tabThrough: false, //allows for tabbing through the different parts of the masked field
 			definitions: {
 				"9": {
 					validator: "[0-9]",
@@ -342,9 +343,9 @@
 			}
 			//test definition => {fn: RegExp/function, cardinality: int, optionality: bool, newBlockMarker: bool, casing: null/upper/lower, def: definitionSymbol, placeholder: placeholder, mask: real maskDefinition}
 			function insertTestDefinition(mtoken, element, position) {
-				var maskdef = opts.definitions[element],
-					newBlockMarker = mtoken.matches.length === 0;
+				var maskdef = opts.definitions[element];
 				position = position !== undefined ? position : mtoken.matches.length;
+				var prevMatch = mtoken.matches[position - 1];
 				if (maskdef && !escaped) {
 					maskdef.placeholder = $.isFunction(maskdef.placeholder) ? maskdef.placeholder.call(this, opts) : maskdef.placeholder;
 					var prevalidators = maskdef.prevalidator,
@@ -359,7 +360,7 @@
 							} : new RegExp("."),
 							cardinality: cardinality ? cardinality : 1,
 							optionality: mtoken.isOptional,
-							newBlockMarker: newBlockMarker,
+							newBlockMarker: prevMatch === undefined || prevMatch.def !== (maskdef.definitionSymbol || element),
 							casing: maskdef.casing,
 							def: maskdef.definitionSymbol || element,
 							placeholder: maskdef.placeholder,
@@ -372,7 +373,7 @@
 						} : new RegExp("."),
 						cardinality: maskdef.cardinality,
 						optionality: mtoken.isOptional,
-						newBlockMarker: newBlockMarker,
+						newBlockMarker: prevMatch === undefined || prevMatch.def !== (maskdef.definitionSymbol || element),
 						casing: maskdef.casing,
 						def: maskdef.definitionSymbol || element,
 						placeholder: maskdef.placeholder,
@@ -383,7 +384,7 @@
 						fn: null,
 						cardinality: 0,
 						optionality: mtoken.isOptional,
-						newBlockMarker: newBlockMarker,
+						newBlockMarker: prevMatch === undefined || prevMatch.def !== element,
 						casing: null,
 						def: element,
 						placeholder: undefined,
@@ -767,13 +768,13 @@
 								posMatch = seekNext(j);
 
 							//does it match
-							if (positionCanMatchDefinition(posMatch, t["match"].def)) {
-								valid = isValid(posMatch, t["input"], true, true) !== false;
+							if (positionCanMatchDefinition(posMatch, t.match.def)) {
+								valid = isValid(posMatch, t.input, true, true) !== false;
 								j = posMatch;
 								break;
 							} else {
-								valid = t["match"].fn == null;
-								if (prevPosMatch == posMatch) break; //prevent endless loop
+								valid = t.match.fn == null;
+								if (prevPosMatch === posMatch) break; //prevent endless loop
 								prevPosMatch = posMatch;
 							}
 						}
@@ -782,37 +783,37 @@
 				}
 
 				if (!valid) {
-					getMaskSet()["validPositions"] = $.extend(true, {}, positionsClone);
+					getMaskSet().validPositions = $.extend(true, {}, positionsClone);
 					return false;
 				}
 			} else
-				getMaskSet()["validPositions"][pos] = validTest;
+				getMaskSet().validPositions[pos] = validTest;
 
 			return true;
 		}
 
 		function stripValidPositions(start, end, nocheck, strict) {
 			var i, startPos = start;
-			getMaskSet()["p"] = start; //needed for alternated position after overtype selection
+			getMaskSet().p = start; //needed for alternated position after overtype selection
 
 			var endPos = end;
 			for (i = startPos; i < end; i++) { //clear selection
-				if (getMaskSet()["validPositions"][i] != undefined) {
-					if (nocheck === true || opts.canClearPosition(getMaskSet(), i, getLastValidPosition(), strict, opts) != false)
-						delete getMaskSet()["validPositions"][i];
+				if (getMaskSet().validPositions[i] !== undefined) {
+					if (nocheck === true || opts.canClearPosition(getMaskSet(), i, getLastValidPosition(), strict, opts) !== false)
+						delete getMaskSet().validPositions[i];
 				}
 			}
 
 			resetMaskSet(true);
 			for (i = startPos + 1; i <= getLastValidPosition();) {
-				while (getMaskSet()["validPositions"][startPos] != undefined) startPos++;
-				var s = getMaskSet()["validPositions"][startPos];
+				while (getMaskSet().validPositions[startPos] !== undefined) startPos++;
+				var s = getMaskSet().validPositions[startPos];
 				if (i < startPos) i = startPos + 1;
-				// while (getMaskSet()["validPositions"][i] == undefined) i++;
-				var t = getMaskSet()["validPositions"][i];
-				if (t != undefined && isMask(i) && s == undefined) {
-					if (positionCanMatchDefinition(startPos, t.match.def) && isValid(startPos, t["input"], true) !== false) {
-						delete getMaskSet()["validPositions"][i];
+				// while (getMaskSet().validPositions[i] == undefined) i++;
+				var t = getMaskSet().validPositions[i];
+				if (t !== undefined && isMask(i) && s === undefined) {
+					if (positionCanMatchDefinition(startPos, t.match.def) && isValid(startPos, t.input, true) !== false) {
+						delete getMaskSet().validPositions[i];
 						i++;
 					}
 					startPos++;
@@ -821,31 +822,31 @@
 			//remove radixpoint if needed
 			var lvp = getLastValidPosition(),
 				ml = getMaskLength();
-			if (strict !== true && nocheck !== true && getMaskSet()["validPositions"][lvp] != undefined && getMaskSet()["validPositions"][lvp].input == opts.radixPoint)
-				delete getMaskSet()["validPositions"][lvp];
+			if (strict !== true && nocheck !== true && getMaskSet().validPositions[lvp] !== undefined && getMaskSet().validPositions[lvp].input === opts.radixPoint)
+				delete getMaskSet().validPositions[lvp];
 
 			for (i = lvp + 1; i <= ml; i++) {
-				if (getMaskSet()["validPositions"][i])
-					delete getMaskSet()["validPositions"][i];
+				if (getMaskSet().validPositions[i])
+					delete getMaskSet().validPositions[i];
 			}
 
 			resetMaskSet(true);
 		}
 
 		function getTestTemplate(pos, ndxIntlzr, tstPs) {
-			var testPos = getMaskSet()["validPositions"][pos];
-			if (testPos == undefined) {
+			var testPos = getMaskSet().validPositions[pos];
+			if (testPos === undefined) {
 				var testPositions = getTests(pos, ndxIntlzr, tstPs),
 					lvp = getLastValidPosition(),
-					lvTest = getMaskSet()["validPositions"][lvp] || getTests(0)[0],
-					lvTestAltArr = (lvTest.alternation != undefined) ? lvTest["locator"][lvTest.alternation].toString().split(',') : [];
+					lvTest = getMaskSet().validPositions[lvp] || getTests(0)[0],
+					lvTestAltArr = (lvTest.alternation !== undefined) ? lvTest.locator[lvTest.alternation].toString().split(",") : [];
 				for (var ndx = 0; ndx < testPositions.length; ndx++) {
 					testPos = testPositions[ndx];
 
-					if (testPos["match"] &&
-						(((opts.greedy && testPos["match"].optionalQuantifier !== true) || (testPos["match"].optionality === false || testPos["match"].newBlockMarker === false) && testPos["match"].optionalQuantifier !== true) &&
-							((lvTest.alternation == undefined || lvTest.alternation != testPos.alternation) ||
-								(testPos["locator"][lvTest.alternation] != undefined && checkAlternationMatch(testPos.locator[lvTest.alternation].toString().split(","), lvTestAltArr))))) {
+					if (testPos.match &&
+						(((opts.greedy && testPos.match.optionalQuantifier !== true) || (testPos.match.optionality === false || testPos.match.newBlockMarker === false) && testPos.match.optionalQuantifier !== true) &&
+							((lvTest.alternation === undefined || lvTest.alternation !== testPos.alternation) ||
+								(testPos.locator[lvTest.alternation] !== undefined && checkAlternationMatch(testPos.locator[lvTest.alternation].toString().split(","), lvTestAltArr))))) {
 						break;
 					}
 				}
@@ -855,53 +856,53 @@
 		}
 
 		function getTest(pos) {
-			if (getMaskSet()['validPositions'][pos]) {
-				return getMaskSet()['validPositions'][pos]["match"];
+			if (getMaskSet().validPositions[pos]) {
+				return getMaskSet().validPositions[pos].match;
 			}
-			return getTests(pos)[0]["match"];
+			return getTests(pos)[0].match;
 		}
 
 		function positionCanMatchDefinition(pos, def) {
 			var valid = false,
 				tests = getTests(pos);
 			for (var tndx = 0; tndx < tests.length; tndx++) {
-				if (tests[tndx]["match"] && tests[tndx]["match"].def == def) {
+				if (tests[tndx].match && tests[tndx].match.def === def) {
 					valid = true;
 					break;
 				}
 			}
 			return valid;
-		};
+		}
 
 		function getTests(pos, ndxIntlzr, tstPs, cacheable) {
-			var maskTokens = getMaskSet()["maskToken"],
+			var maskTokens = getMaskSet().maskToken,
 				testPos = ndxIntlzr ? tstPs : 0,
 				ndxInitializer = ndxIntlzr || [0],
 				matches = [],
 				insertStop = false;
 
-			function ResolveTestFromToken(maskToken, ndxInitializer, loopNdx, quantifierRecurse) { //ndxInitilizer contains a set of indexes to speedup searches in the mtokens
+			function resolveTestFromToken(maskToken, ndxInitializer, loopNdx, quantifierRecurse) { //ndxInitilizer contains a set of indexes to speedup searches in the mtokens
 				function handleMatch(match, loopNdx, quantifierRecurse) {
 					if (testPos > 10000) {
-						alert("jquery.inputmask: There is probably an error in your mask definition or in the code. Create an issue on github with an example of the mask you are using. " + getMaskSet()["mask"]);
+						alert("jquery.inputmask: There is probably an error in your mask definition or in the code. Create an issue on github with an example of the mask you are using. " + getMaskSet().mask);
 						return true;
 					}
-					if (testPos == pos && match.matches == undefined) {
+					if (testPos === pos && match.matches === undefined) {
 						matches.push({
 							"match": match,
 							"locator": loopNdx.reverse()
 						});
 						return true;
-					} else if (match.matches != undefined) {
+					} else if (match.matches !== undefined) {
 						if (match.isGroup && quantifierRecurse !== match) { //when a group pass along to the quantifier
 							match = handleMatch(maskToken.matches[tndx + 1], loopNdx);
 							if (match) return true;
 						} else if (match.isOptional) {
 							var optionalToken = match;
-							match = ResolveTestFromToken(match, ndxInitializer, loopNdx, quantifierRecurse);
+							match = resolveTestFromToken(match, ndxInitializer, loopNdx, quantifierRecurse);
 							if (match) {
-								var latestMatch = matches[matches.length - 1]["match"];
-								var isFirstMatch = $.inArray(latestMatch, optionalToken.matches) == 0;
+								var latestMatch = matches[matches.length - 1].match;
+								var isFirstMatch = $.inArray(latestMatch, optionalToken.matches) === 0;
 								if (isFirstMatch) {
 									insertStop = true; //insert a stop
 									testPos = pos; //match the position after the group
@@ -914,7 +915,7 @@
 								currentMatches = matches.slice(),
 								loopNdxCnt = loopNdx.length;
 							var altIndex = ndxInitializer.length > 0 ? ndxInitializer.shift() : -1;
-							if (altIndex == -1 || typeof altIndex == "string") {
+							if (altIndex === -1 || typeof altIndex === "string") {
 								var currentPos = testPos,
 									ndxInitializerClone = ndxInitializer.slice(),
 									altIndexArr = [];
@@ -922,10 +923,10 @@
 								for (var amndx = 0; amndx < alternateToken.matches.length; amndx++) {
 									matches = [];
 									match = handleMatch(alternateToken.matches[amndx], [amndx].concat(loopNdx), quantifierRecurse) || match;
-									if (match !== true && match != undefined && (altIndexArr[altIndexArr.length - 1] < alternateToken.matches.length)) { //no match in the alternations (length mismatch) => look further
+									if (match !== true && match !== undefined && (altIndexArr[altIndexArr.length - 1] < alternateToken.matches.length)) { //no match in the alternations (length mismatch) => look further
 										var ntndx = maskToken.matches.indexOf(match) + 1;
 										if (maskToken.matches.length > ntndx) {
-											match = handleMatch(maskToken.matches[ntndx], [ntndx].concat(loopNdx.slice(1, loopNdx.length)), quantifierRecurse)
+											match = handleMatch(maskToken.matches[ntndx], [ntndx].concat(loopNdx.slice(1, loopNdx.length)), quantifierRecurse);
 											if (match) {
 												altIndexArr.push(ntndx.toString());
 												$.each(matches, function(ndx, lmnt) {
@@ -948,7 +949,7 @@
 										for (var ndx2 = 0; ndx2 < malternateMatches.length; ndx2++) {
 											var altMatch2 = malternateMatches[ndx2];
 											//verify equality
-											if (altMatch.match.mask == altMatch2.match.mask && (typeof altIndex != "string" || $.inArray(altMatch.locator[altMatch.alternation].toString(), altIndexArr) != -1)) {
+											if (altMatch.match.mask === altMatch2.match.mask && (typeof altIndex !== "string" || $.inArray(altMatch.locator[altMatch.alternation].toString(), altIndexArr) !== -1)) {
 												maltMatches.splice(ndx1, 1);
 												ndx1--;
 												altMatch2.locator[altMatch.alternation] = altMatch2.locator[altMatch.alternation] + "," + altMatch.locator[altMatch.alternation];
@@ -969,9 +970,9 @@
 											lmnt.locator[alternation] = undefined;
 											lmnt.alternation = undefined;
 											for (var alndx = 0; alndx < altLocArr.length; alndx++) {
-												mamatch = $.inArray(altLocArr[alndx], altIndexArr) != -1;
+												mamatch = $.inArray(altLocArr[alndx], altIndexArr) !== -1;
 												if (mamatch) { //rebuild the locator with valid entries
-													if (lmnt.locator[alternation] != undefined) {
+													if (lmnt.locator[alternation] !== undefined) {
 														lmnt.locator[alternation] += ",";
 														lmnt.locator[alternation] += altLocArr[alndx];
 													} else
@@ -980,7 +981,7 @@
 													lmnt.alternation = alternation;
 												}
 											}
-											if (lmnt.locator[alternation] != undefined) return lmnt;
+											if (lmnt.locator[alternation] !== undefined) return lmnt;
 										}
 									});
 								}
@@ -1002,9 +1003,9 @@
 								match = handleMatch(tokenGroup, [qndx].concat(loopNdx), tokenGroup); //set the tokenGroup as quantifierRecurse marker
 								if (match) {
 									//get latest match
-									var latestMatch = matches[matches.length - 1]["match"];
+									var latestMatch = matches[matches.length - 1].match;
 									latestMatch.optionalQuantifier = qndx > (qt.quantifier.min - 1);
-									var isFirstMatch = $.inArray(latestMatch, tokenGroup.matches) == 0;
+									var isFirstMatch = $.inArray(latestMatch, tokenGroup.matches) === 0;
 
 									if (isFirstMatch) { //search for next possible match
 										if (qndx > (qt.quantifier.min - 1)) {
@@ -1018,7 +1019,7 @@
 								}
 							}
 						} else {
-							match = ResolveTestFromToken(match, ndxInitializer, loopNdx, quantifierRecurse);
+							match = resolveTestFromToken(match, ndxInitializer, loopNdx, quantifierRecurse);
 							if (match)
 								return true;
 						}
@@ -1026,9 +1027,9 @@
 				}
 
 				for (var tndx = (ndxInitializer.length > 0 ? ndxInitializer.shift() : 0); tndx < maskToken.matches.length; tndx++) {
-					if (maskToken.matches[tndx]["isQuantifier"] !== true) {
+					if (maskToken.matches[tndx].isQuantifier !== true) {
 						var match = handleMatch(maskToken.matches[tndx], [tndx].concat(loopNdx), quantifierRecurse);
-						if (match && testPos == pos) {
+						if (match && testPos === pos) {
 							return match;
 						} else if (testPos > pos) {
 							break;
@@ -1037,30 +1038,30 @@
 				}
 			}
 
-			if (cacheable === true && getMaskSet()['tests'][pos]) {
-				return getMaskSet()['tests'][pos];
+			if (cacheable === true && getMaskSet().tests[pos]) {
+				return getMaskSet().tests[pos];
 			}
-			if (ndxIntlzr == undefined) {
+			if (ndxIntlzr === undefined) {
 				var previousPos = pos - 1,
 					test;
-				while ((test = getMaskSet()['validPositions'][previousPos]) == undefined && previousPos > -1) {
-					if (getMaskSet()['tests'][previousPos] && (test = getMaskSet()['tests'][previousPos][0]) != undefined)
+				while ((test = getMaskSet().validPositions[previousPos]) === undefined && previousPos > -1) {
+					if (getMaskSet().tests[previousPos] && (test = getMaskSet().tests[previousPos][0]) !== undefined)
 						break;
 					previousPos--;
 
 				}
-				if (test != undefined && previousPos > -1) {
+				if (test !== undefined && previousPos > -1) {
 					testPos = previousPos;
-					ndxInitializer = test["locator"].slice();
+					ndxInitializer = test.locator.slice();
 				}
 			}
 			for (var mtndx = ndxInitializer.shift(); mtndx < maskTokens.length; mtndx++) {
-				var match = ResolveTestFromToken(maskTokens[mtndx], ndxInitializer, [mtndx]);
-				if ((match && testPos == pos) || testPos > pos) {
+				var match = resolveTestFromToken(maskTokens[mtndx], ndxInitializer, [mtndx]);
+				if ((match && testPos === pos) || testPos > pos) {
 					break;
 				}
 			}
-			if (matches.length == 0 || insertStop)
+			if (matches.length === 0 || insertStop)
 				matches.push({
 					"match": {
 						fn: null,
@@ -1072,10 +1073,10 @@
 					"locator": []
 				});
 
-			getMaskSet()['tests'][pos] = $.extend(true, [], matches); //set a clone to prevent overwriting some props
+			getMaskSet().tests[pos] = $.extend(true, [], matches); //set a clone to prevent overwriting some props
 
 			// console.log(pos + " - " + JSON.stringify(matches));
-			return getMaskSet()['tests'][pos];
+			return getMaskSet().tests[pos];
 		}
 
 		function getBufferTemplate() {
@@ -1090,7 +1091,7 @@
 			if (getMaskSet().buffer === undefined) {
 				getMaskSet().buffer = getMaskTemplate(true, getLastValidPosition(), true);
 			}
-			return getMaskSet()['buffer'];
+			return getMaskSet().buffer;
 		}
 
 		function refreshFromBuffer(start, end, buffer) {
@@ -1101,13 +1102,13 @@
 				end = buffer.length;
 			} else {
 				for (var i = start; i < end; i++) {
-					delete getMaskSet()["validPositions"][i];
-					delete getMaskSet()["tests"][i];
+					delete getMaskSet().validPositions[i];
+					delete getMaskSet().tests[i];
 				}
 			}
 
 			for (var i = start; i < end; i++) {
-				if (buffer[i] != opts.skipOptionalPartCharacter) {
+				if (buffer[i] !== opts.skipOptionalPartCharacter) {
 					isValid(i, buffer[i], true, true);
 				}
 			}
@@ -1145,7 +1146,7 @@
 				var rslt = false;
 				//console.log(JSON.stringify(getTests(position)));
 				$.each(getTests(position), function(ndx, tst) {
-					var test = tst["match"];
+					var test = tst.match;
 					var loopend = c ? 1 : 0,
 						chrs = '',
 						buffer = getBuffer();
@@ -1158,15 +1159,15 @@
 
 					//return is false or a json object => { pos: ??, c: ??} or true
 					rslt = test.fn != null ?
-						test.fn.test(chrs, getMaskSet(), position, strict, opts) : (c == test["def"] || c == opts.skipOptionalPartCharacter) && test["def"] != "" ? //non mask
+						test.fn.test(chrs, getMaskSet(), position, strict, opts) : (c == test.def || c == opts.skipOptionalPartCharacter) && test.def != "" ? //non mask
 						{
-							c: test["def"],
+							c: test.def,
 							pos: position
 						} : false;
 
 					if (rslt !== false) {
 						var elem = rslt.c != undefined ? rslt.c : c;
-						elem = (elem == opts.skipOptionalPartCharacter && test["fn"] === null) ? test["def"] : elem;
+						elem = (elem == opts.skipOptionalPartCharacter && test.fn === null) ? test.def : elem;
 
 						var validatedPos = position,
 							possibleModifiedBuffer = getBuffer();
@@ -1179,39 +1180,39 @@
 								stripValidPositions(lmnt, lmnt + 1, true);
 							});
 						}
-						if (rslt["insert"] != undefined) { //insert position(s)
-							if (!$.isArray(rslt["insert"])) rslt["insert"] = [rslt["insert"]];
-							$.each(rslt["insert"].sort(function(a, b) {
+						if (rslt.insert != undefined) { //insert position(s)
+							if (!$.isArray(rslt.insert)) rslt.insert = [rslt.insert];
+							$.each(rslt.insert.sort(function(a, b) {
 								return a - b;
 							}), function(ndx, lmnt) {
-								isValid(lmnt["pos"], lmnt["c"], true);
+								isValid(lmnt.pos, lmnt.c, true);
 							});
 						}
 
-						if (rslt["refreshFromBuffer"]) {
-							var refresh = rslt["refreshFromBuffer"];
+						if (rslt.refreshFromBuffer) {
+							var refresh = rslt.refreshFromBuffer;
 							strict = true;
-							refreshFromBuffer(refresh === true ? refresh : refresh["start"], refresh["end"], possibleModifiedBuffer);
-							if (rslt.pos == undefined && rslt.c == undefined) {
+							refreshFromBuffer(refresh === true ? refresh : refresh.start, refresh.end, possibleModifiedBuffer);
+							if (rslt.pos === undefined && rslt.c === undefined) {
 								rslt.pos = getLastValidPosition();
 								return false; //breakout if refreshFromBuffer && nothing to insert
 							}
-							validatedPos = rslt.pos != undefined ? rslt.pos : position;
+							validatedPos = rslt.pos !== undefined ? rslt.pos : position;
 							if (validatedPos != position) {
 								rslt = $.extend(rslt, isValid(validatedPos, elem, true)); //revalidate new position strict
 								return false;
 							}
 
-						} else if (rslt !== true && rslt.pos != undefined && rslt["pos"] != position) { //their is a position offset
-							validatedPos = rslt["pos"];
+						} else if (rslt !== true && rslt.pos !== undefined && rslt.pos !== position) { //their is a position offset
+							validatedPos = rslt.pos;
 							refreshFromBuffer(position, validatedPos);
-							if (validatedPos != position) {
+							if (validatedPos !== position) {
 								rslt = $.extend(rslt, isValid(validatedPos, elem, true)); //revalidate new position strict
 								return false;
 							}
 						}
 
-						if (rslt != true && rslt.pos == undefined && rslt.c == undefined) {
+						if (rslt !== true && rslt.pos === undefined && rslt.c === undefined) {
 							return false; //breakout if nothing to insert
 						}
 
@@ -1231,30 +1232,30 @@
 			}
 
 			function alternate(pos, c, strict, fromSetValid) {
-				var validPsClone = $.extend(true, {}, getMaskSet()["validPositions"]),
+				var validPsClone = $.extend(true, {}, getMaskSet().validPositions),
 					lastAlt,
 					alternation,
 					isValidRslt,
 					altPos;
 				//find last modified alternation
 				for (var lAlt = getLastValidPosition(); lAlt >= 0; lAlt--) {
-					altPos = getMaskSet()["validPositions"][lAlt];
-					if (altPos && altPos.alternation != undefined) {
+					altPos = getMaskSet().validPositions[lAlt];
+					if (altPos && altPos.alternation !== undefined) {
 						lastAlt = lAlt;
-						alternation = getMaskSet()["validPositions"][lastAlt].alternation;
-						if (getTestTemplate(lastAlt).locator[altPos.alternation] != altPos.locator[altPos.alternation]) {
+						alternation = getMaskSet().validPositions[lastAlt].alternation;
+						if (getTestTemplate(lastAlt).locator[altPos.alternation] !== altPos.locator[altPos.alternation]) {
 							break;
 						}
 					}
 				}
-				if (alternation != undefined) {
+				if (alternation !== undefined) {
 					//find first decision making position
 					lastAlt = parseInt(lastAlt);
-					for (var decisionPos in getMaskSet()["validPositions"]) {
+					for (var decisionPos in getMaskSet().validPositions) {
 						decisionPos = parseInt(decisionPos);
-						altPos = getMaskSet()["validPositions"][decisionPos];
-						if (decisionPos >= lastAlt && altPos.alternation != undefined) {
-							var altNdxs = getMaskSet()["validPositions"][lastAlt].locator[alternation].toString().split(','),
+						altPos = getMaskSet().validPositions[decisionPos];
+						if (decisionPos >= lastAlt && altPos.alternation !== undefined) {
+							var altNdxs = getMaskSet().validPositions[lastAlt].locator[alternation].toString().split(','),
 								decisionTaker = altPos.locator[alternation] || altNdxs[0]; //no match in the alternations (length mismatch)
 							if (decisionTaker.length > 0) { //no decision taken ~ take first one as decider
 								decisionTaker = decisionTaker.split(',')[0];
@@ -1264,26 +1265,26 @@
 								if (decisionTaker < altNdxs[mndx]) {
 									var possibilityPos, possibilities;
 									for (var dp = decisionPos; dp >= 0; dp--) {
-										possibilityPos = getMaskSet()["validPositions"][dp];
-										if (possibilityPos != undefined) {
+										possibilityPos = getMaskSet().validPositions[dp];
+										if (possibilityPos !== undefined) {
 											possibilities = possibilityPos.locator[alternation]; //store to reset
 											//possibilityPos.alternation = undefined;
 											possibilityPos.locator[alternation] = parseInt(altNdxs[mndx]);
 											break;
 										}
 									}
-									if (decisionTaker != possibilityPos.locator[alternation]) {
+									if (decisionTaker !== possibilityPos.locator[alternation]) {
 										var validInputs = [],
 											staticInputsBeforePos = 0;
 										for (var i = decisionPos + 1; i < getLastValidPosition() + 1; i++) {
-											var validPos = getMaskSet()["validPositions"][i];
+											var validPos = getMaskSet().validPositions[i];
 											if (validPos) {
 												if (validPos.match.fn != null) {
 													validInputs.push(validPos.input);
 												} else if (i < pos) staticInputsBeforePos++;
 											}
-											delete getMaskSet()["validPositions"][i];
-											delete getMaskSet()["tests"][i];
+											delete getMaskSet().validPositions[i];
+											delete getMaskSet().tests[i];
 										}
 										resetMaskSet(true); //clear getbuffer
 										opts.keepStatic = !opts.keepStatic; //disable keepStatic on getMaskLength
@@ -1291,7 +1292,7 @@
 										while (validInputs.length > 0) {
 											var input = validInputs.shift();
 											//console.log(input);
-											if (input != opts.skipOptionalPartCharacter) {
+											if (input !== opts.skipOptionalPartCharacter) {
 												if (!(isValidRslt = isValid(getLastValidPosition() + 1, input, false, true))) {
 													break;
 												}
@@ -1304,7 +1305,7 @@
 											var targetLvp = getLastValidPosition(pos) + 1;
 											var staticInputsBeforePosAlternate = 0;
 											for (var i = decisionPos + 1; i < getLastValidPosition() + 1; i++) {
-												var validPos = getMaskSet()["validPositions"][i];
+												var validPos = getMaskSet().validPositions[i];
 												if (validPos && validPos.match.fn == null && i < pos)
 													staticInputsBeforePosAlternate++;
 											}
@@ -1314,7 +1315,7 @@
 										opts.keepStatic = !opts.keepStatic; //enable keepStatic on getMaskLength
 										if (!isValidRslt) {
 											resetMaskSet();
-											getMaskSet()["validPositions"] = $.extend(true, {}, validPsClone);
+											getMaskSet().validPositions = $.extend(true, {}, validPsClone);
 										} else
 											return isValidRslt;
 									}
@@ -1329,7 +1330,7 @@
 
 			//set alternator choice on previous skipped placeholder positions
 			function trackbackAlternations(originalPos, newPos) {
-				var vp = getMaskSet()["validPositions"][newPos],
+				var vp = getMaskSet().validPositions[newPos],
 					targetLocator = vp.locator,
 					tll = targetLocator.length;
 
@@ -1347,8 +1348,8 @@
 							}
 						});
 						setValidPosition(ps, $.extend({}, bestMatch, {
-							"input": bestMatch["match"].def
-						}), true)
+							"input": bestMatch.match.def
+						}), true);
 					}
 				}
 			}
@@ -1356,14 +1357,14 @@
 			var buffer = getBuffer();
 			//find previous valid
 			for (var pndx = pos - 1; pndx > -1; pndx--) {
-				if (getMaskSet()["validPositions"][pndx])
+				if (getMaskSet().validPositions[pndx])
 					break;
 			}
 			////fill missing nonmask and valid placeholders
 			pndx++;
 			for (; pndx < pos; pndx++) {
 				//console.log("missing " + pndx + " " + buffer[pndx] + " ismask " + isMask(pndx) + " plchldr " + getPlaceholder(pndx) + " nrt " + getTests(pndx).len);
-				if (getMaskSet()["validPositions"][pndx] == undefined && (((!isMask(pndx) || buffer[pndx] != getPlaceholder(pndx)) && getTests(pndx).length > 1) || (buffer[pndx] == opts.radixPoint || buffer[pndx] == "0" && $.inArray(opts.radixPoint, buffer) < pndx))) //special case for decimals ~ = placeholder but yet valid input
+				if (getMaskSet().validPositions[pndx] === undefined && (((!isMask(pndx) || buffer[pndx] !== getPlaceholder(pndx)) && getTests(pndx).length > 1) || (buffer[pndx] === opts.radixPoint || buffer[pndx] === "0" && $.inArray(opts.radixPoint, buffer) < pndx))) //special case for decimals ~ = placeholder but yet valid input
 				{
 					//console.log("inject " + pndx + " " + buffer[pndx]);
 					_isValid(pndx, buffer[pndx], true);
@@ -1372,7 +1373,7 @@
 
 			var maskPos = pos,
 				result = false,
-				positionsClone = $.extend(true, {}, getMaskSet()["validPositions"]); //clone the currentPositions
+				positionsClone = $.extend(true, {}, getMaskSet().validPositions); //clone the currentPositions
 
 			//if (fromSetValid && maskPos >= getMaskLength()) {
 			//		resetMaskSet(true); //masklenght can be altered on the process => reset to get the actual length
@@ -1380,12 +1381,12 @@
 			if (maskPos < getMaskLength()) {
 				result = _isValid(maskPos, c, strict, fromSetValid);
 				if ((!strict || fromSetValid) && result === false) {
-					var currentPosValid = getMaskSet()["validPositions"][maskPos];
-					if (currentPosValid && currentPosValid["match"].fn == null && (currentPosValid["match"].def == c || c == opts.skipOptionalPartCharacter)) {
+					var currentPosValid = getMaskSet().validPositions[maskPos];
+					if (currentPosValid && currentPosValid.match.fn === null && (currentPosValid.match.def === c || c === opts.skipOptionalPartCharacter)) {
 						result = {
 							"caret": seekNext(maskPos)
 						};
-					} else if ((opts.insertMode || getMaskSet()["validPositions"][seekNext(maskPos)] == undefined) && !isMask(maskPos)) { //does the input match on a further position?
+					} else if ((opts.insertMode || getMaskSet().validPositions[seekNext(maskPos)] === undefined) && !isMask(maskPos)) { //does the input match on a further position?
 						for (var nPos = maskPos + 1, snPos = seekNext(maskPos); nPos <= snPos; nPos++) {
 							result = _isValid(nPos, c, strict, fromSetValid);
 							if (result !== false) {
@@ -1404,16 +1405,16 @@
 				"pos": maskPos
 			};
 
-			if ($.isFunction(opts.postValidation) && result != false && !strict) {
+			if ($.isFunction(opts.postValidation) && result !== false && !strict) {
 				resetMaskSet(true);
 				var postValidResult = opts.postValidation(getBuffer(), opts);
 				if (!postValidResult) {
 					resetMaskSet(true);
-					getMaskSet()["validPositions"] = $.extend(true, {}, positionsClone); //revert validation changes
+					getMaskSet().validPositions = $.extend(true, {}, positionsClone); //revert validation changes
 					result = false;
-				} else if (postValidResult["refreshFromBuffer"]) {
-					var refresh = postValidResult["refreshFromBuffer"];
-					refreshFromBuffer(refresh === true ? refresh : refresh["start"], refresh["end"], postValidResult["buffer"]);
+				} else if (postValidResult.refreshFromBuffer) {
+					var refresh = postValidResult.refreshFromBuffer;
+					refreshFromBuffer(refresh === true ? refresh : refresh.start, refresh.end, postValidResult.buffer);
 					resetMaskSet(true);
 					result = postValidResult;
 				}
@@ -1425,12 +1426,12 @@
 		function isMask(pos) {
 			var test = getTest(pos);
 			if (test.fn != null)
-				return test.fn
-			else if (!opts.keepStatic && getMaskSet()["validPositions"][pos] == undefined) {
+				return test.fn;
+			else if (!opts.keepStatic && getMaskSet().validPositions[pos] === undefined) {
 				var tests = getTests(pos),
 					staticAlternations = true;
 				for (var i = 0; i < tests.length; i++) {
-					if (tests[i].match.def != "" && ( /*tests[i].match.fn !== null || */ (tests[i].alternation == undefined || tests[i].locator[tests[i].alternation].length > 1))) {
+					if (tests[i].match.def !== "" && ( /*tests[i].match.fn !== null || */ (tests[i].alternation === undefined || tests[i].locator[tests[i].alternation].length > 1))) {
 						staticAlternations = false;
 						break;
 					}
@@ -1443,57 +1444,57 @@
 
 		function getMaskLength() {
 			var maskLength;
-			maxLength = $el.prop('maxLength');
-			if (maxLength == -1) maxLength = undefined; /* FF sets no defined max length to -1 */
+			maxLength = $el.prop("maxLength");
+			if (maxLength === -1) maxLength = undefined; /* FF sets no defined max length to -1 */
 			var pos, lvp = getLastValidPosition(),
-				testPos = getMaskSet()["validPositions"][lvp],
-				ndxIntlzr = testPos != undefined ? testPos["locator"].slice() : undefined;
-			for (pos = lvp + 1; testPos == undefined || (testPos["match"]["fn"] != null || (testPos["match"]["fn"] == null && testPos["match"]["def"] != "")); pos++) {
+				testPos = getMaskSet().validPositions[lvp],
+				ndxIntlzr = testPos !== undefined ? testPos.locator.slice() : undefined;
+			for (pos = lvp + 1; testPos === undefined || (testPos.match.fn !== null || (testPos.match.fn === null && testPos.match.def !== "")); pos++) {
 				testPos = getTestTemplate(pos, ndxIntlzr, pos - 1);
-				ndxIntlzr = testPos["locator"].slice();
+				ndxIntlzr = testPos.locator.slice();
 			}
 
 			var lastTest = getTest(pos - 1);
-			maskLength = (lastTest.def != "") ? pos : pos - 1;
-			return (maxLength == undefined || maskLength < maxLength) ? maskLength : maxLength;
+			maskLength = (lastTest.def !== "") ? pos : pos - 1;
+			return (maxLength === undefined || maskLength < maxLength) ? maskLength : maxLength;
 		}
 
-		function seekNext(pos) {
+		function seekNext(pos, newBlock) {
 			var maskL = getMaskLength();
 			if (pos >= maskL) return maskL;
 			var position = pos;
-			while (++position < maskL && !isMask(position) && (opts.nojumps !== true || opts.nojumpsThreshold > position)) {}
-
+			while (++position < maskL && ((newBlock === true && getTest(position).newBlockMarker !== true || !isMask(position)) || (newBlock !== true && !isMask(position) && (opts.nojumps !== true || opts.nojumpsThreshold > position)))) {}
 			return position;
 		}
 
-		function seekPrevious(pos) {
+		function seekPrevious(pos, newBlock) {
 			var position = pos;
 			if (position <= 0) return 0;
 
-			while (--position > 0 && !isMask(position)) {};
+			while (--position > 0 && ((newBlock === true && getTest(position).newBlockMarker !== true) || (newBlock !== true && !isMask(position)))) {}
+
 			return position;
 		}
 
 		function getBufferElement(position) {
-			return getMaskSet()["validPositions"][position] == undefined ? getPlaceholder(position) : getMaskSet()["validPositions"][position]["input"];
+			return getMaskSet().validPositions[position] === undefined ? getPlaceholder(position) : getMaskSet().validPositions[position].input;
 		}
 
 		function writeBuffer(input, buffer, caretPos, event, triggerInputEvent) {
 			if (event && $.isFunction(opts.onBeforeWrite)) {
 				var result = opts.onBeforeWrite.call(input, event, buffer, caretPos, opts);
 				if (result) {
-					if (result["refreshFromBuffer"]) {
-						var refresh = result["refreshFromBuffer"];
-						refreshFromBuffer(refresh === true ? refresh : refresh["start"], refresh["end"], result["buffer"] || buffer);
+					if (result.refreshFromBuffer) {
+						var refresh = result.refreshFromBuffer;
+						refreshFromBuffer(refresh === true ? refresh : refresh.start, refresh.end, result.buffer || buffer);
 						resetMaskSet(true);
 						buffer = getBuffer();
 					}
-					caretPos = result.caret != undefined ? result.caret : caretPos;
+					caretPos = result.caret !== undefined ? result.caret : caretPos;
 				}
 			}
 			input.inputmask._valueSet(buffer.join(''));
-			if (caretPos != undefined) {
+			if (caretPos !== undefined) {
 				caret(input, caretPos);
 			}
 			if (triggerInputEvent === true) {
@@ -1504,27 +1505,27 @@
 
 		function getPlaceholder(pos, test) {
 			test = test || getTest(pos);
-			if (test["placeholder"] != undefined)
-				return test["placeholder"];
-			else if (test["fn"] == null) {
-				if (!opts.keepStatic && getMaskSet()["validPositions"][pos] == undefined) {
+			if (test.placeholder !== undefined)
+				return test.placeholder;
+			else if (test.fn === null) {
+				if (!opts.keepStatic && getMaskSet().validPositions[pos] === undefined) {
 					var tests = getTests(pos),
 						hasAlternations = false,
 						prevTest;
 					for (var i = 0; i < tests.length; i++) {
-						if (prevTest && tests[i].match.def != "" && (tests[i].match.def != prevTest.match.def && (tests[i].alternation == undefined || tests[i].alternation == prevTest.alternation))) {
+						if (prevTest && tests[i].match.def !== "" && (tests[i].match.def !== prevTest.match.def && (tests[i].alternation === undefined || tests[i].alternation === prevTest.alternation))) {
 							hasAlternations = true;
 							break;
 						}
 
-						if (tests[i].match.optionality != true && tests[i].match.optionalQuantifier != true)
+						if (tests[i].match.optionality !== true && tests[i].match.optionalQuantifier !== true)
 							prevTest = tests[i];
 					}
 
 					if (hasAlternations)
 						return opts.placeholder.charAt(pos % opts.placeholder.length);
 				}
-				return test["def"]
+				return test.def;
 			} else {
 				return opts.placeholder.charAt(pos % opts.placeholder.length);
 			}
@@ -1534,11 +1535,11 @@
 			function isTemplateMatch() {
 				var isMatch = false;
 				var charCodeNdx = getBufferTemplate().slice(initialNdx, seekNext(initialNdx)).join('').indexOf(charCodes);
-				if (charCodeNdx != -1 && !isMask(initialNdx)) {
+				if (charCodeNdx !== -1 && !isMask(initialNdx)) {
 					isMatch = true;
 					var bufferTemplateArr = getBufferTemplate().slice(initialNdx, initialNdx + charCodeNdx);
 					for (var i = 0; i < bufferTemplateArr.length; i++) {
-						if (bufferTemplateArr[i] != " ") {
+						if (bufferTemplateArr[i] !== " ") {
 							isMatch = false;
 							break;
 						}
@@ -1547,15 +1548,15 @@
 
 				return isMatch;
 			}
-			var inputValue = nptvl != undefined ? nptvl.slice() : input.inputmask._valueGet().split(''),
+			var inputValue = nptvl !== undefined ? nptvl.slice() : input.inputmask._valueGet().split(''),
 				charCodes = "",
 				initialNdx = 0;
 			resetMaskSet();
-			getMaskSet()["p"] = seekNext(-1);
+			getMaskSet().p = seekNext(-1);
 			if (writeOut) input.inputmask._valueSet(""); //initial clear
 
 			if (!strict) {
-				if (opts.autoUnmask != true) {
+				if (opts.autoUnmask !== true) {
 					var staticInput = getBufferTemplate().slice(0, seekNext(-1)).join(''),
 						matches = inputValue.join('').match(new RegExp("^" + Inputmask.escapeRegex(staticInput), "g"));
 					if (matches && matches.length > 0) {
@@ -1573,10 +1574,10 @@
 				keypress.which = charCode.charCodeAt(0);
 				charCodes += charCode;
 				var lvp = getLastValidPosition(undefined, true),
-					lvTest = getMaskSet()["validPositions"][lvp],
+					lvTest = getMaskSet().validPositions[lvp],
 					nextTest = getTestTemplate(lvp + 1, lvTest ? lvTest.locator.slice() : undefined, lvp);
 				if (!isTemplateMatch() || strict || opts.autoUnmask) {
-					var pos = strict ? ndx : (nextTest["match"].fn == null && nextTest["match"].optionality && (lvp + 1) < getMaskSet()["p"] ? lvp + 1 : getMaskSet()["p"]);
+					var pos = strict ? ndx : (nextTest.match.fn == null && nextTest.match.optionality && (lvp + 1) < getMaskSet().p ? lvp + 1 : getMaskSet().p);
 					keypressEvent.call(input, keypress, true, false, strict, pos);
 					initialNdx = pos + 1;
 					charCodes = "";
@@ -1757,7 +1758,7 @@
 								if (imOpts)
 									(new Inputmask(imOpts)).mask(this);
 								else $(this).unbind(".inputmask");
-							} else if (e.type !== "setvalue" && (this.disabled || (this.readOnly && !(e.type === "keydown" && (e.ctrlKey && e.keyCode === 67) || e.keyCode === Inputmask.keyCode.TAB))))
+							} else if (e.type !== "setvalue" && (this.disabled || (this.readOnly && !(e.type === "keydown" && (e.ctrlKey && e.keyCode === 67) || (opts.tabThrough === false && e.keyCode === Inputmask.keyCode.TAB)))))
 								e.preventDefault();
 							else {
 								switch (e.type) {
@@ -1972,12 +1973,12 @@
 
 			if (k === Inputmask.keyCode.BACKSPACE && (pos.end - pos.begin < 1 || opts.insertMode === false)) {
 				pos.begin = seekPrevious(pos.begin);
-				if (getMaskSet()["validPositions"][pos.begin] !== undefined && (getMaskSet()["validPositions"][pos.begin].input === opts.groupSeparator || getMaskSet()["validPositions"][pos.begin].input === opts.radixPoint)) {
+				if (getMaskSet().validPositions[pos.begin] !== undefined && (getMaskSet().validPositions[pos.begin].input === opts.groupSeparator || getMaskSet().validPositions[pos.begin].input === opts.radixPoint)) {
 					pos.begin--;
 				}
 			} else if (k == Inputmask.keyCode.DELETE && pos.begin == pos.end) {
 				pos.end = isMask(pos.end) ? pos.end + 1 : seekNext(pos.end) + 1;
-				if (getMaskSet()["validPositions"][pos.begin] != undefined && (getMaskSet()["validPositions"][pos.begin].input == opts.groupSeparator || getMaskSet()["validPositions"][pos.begin].input == opts.radixPoint)) {
+				if (getMaskSet().validPositions[pos.begin] != undefined && (getMaskSet().validPositions[pos.begin].input == opts.groupSeparator || getMaskSet().validPositions[pos.begin].input == opts.radixPoint)) {
 					pos.end++;
 				}
 			}
@@ -1989,9 +1990,9 @@
 			var lvp = getLastValidPosition(pos.begin);
 			if (lvp < pos.begin) {
 				if (lvp == -1) resetMaskSet();
-				getMaskSet()["p"] = seekNext(lvp);
+				getMaskSet().p = seekNext(lvp);
 			} else if (strict !== true) {
-				getMaskSet()["p"] = pos.begin;
+				getMaskSet().p = pos.begin;
 			}
 		}
 
@@ -2002,39 +2003,54 @@
 				pos = caret(input);
 
 			//backspace, delete, and escape get special treatment
-			if (k == Inputmask.keyCode.BACKSPACE || k == Inputmask.keyCode.DELETE || (iphone && k == 127) || (e.ctrlKey && k == 88 && !isInputEventSupported("cut"))) { //backspace/delete
+			if (k === Inputmask.keyCode.BACKSPACE || k === Inputmask.keyCode.DELETE || (iphone && k === 127) || (e.ctrlKey && k === 88 && !isInputEventSupported("cut"))) { //backspace/delete
 				e.preventDefault(); //stop default action but allow propagation
 				if (k == 88) undoValue = getBuffer().join('');
 				handleRemove(input, k, pos);
-				writeBuffer(input, getBuffer(), getMaskSet()["p"], e, undoValue != getBuffer().join(''));
+				writeBuffer(input, getBuffer(), getMaskSet().p, e, undoValue != getBuffer().join(''));
 				if (input.inputmask._valueGet() == getBufferTemplate().join(''))
 					$input.trigger('cleared');
 				else if (isComplete(getBuffer()) === true)
 					$input.trigger("complete");
 				if (opts.showTooltip) { //update tooltip
-					$input.prop("title", getMaskSet()["mask"]);
+					$input.prop("title", getMaskSet().mask);
 				}
-			} else if (k == Inputmask.keyCode.END || k == Inputmask.keyCode.PAGE_DOWN) { //when END or PAGE_DOWN pressed set position at lastmatch
+			} else if (k === Inputmask.keyCode.END || k === Inputmask.keyCode.PAGE_DOWN) { //when END or PAGE_DOWN pressed set position at lastmatch
 				setTimeout(function() {
 					var caretPos = seekNext(getLastValidPosition());
-					if (!opts.insertMode && caretPos == getMaskLength() && !e.shiftKey) caretPos--;
+					if (!opts.insertMode && caretPos === getMaskLength() && !e.shiftKey) caretPos--;
 					caret(input, e.shiftKey ? pos.begin : caretPos, caretPos);
 				}, 0);
-			} else if ((k == Inputmask.keyCode.HOME && !e.shiftKey) || k == Inputmask.keyCode.PAGE_UP) { //Home or page_up
+			} else if ((k === Inputmask.keyCode.HOME && !e.shiftKey) || k === Inputmask.keyCode.PAGE_UP) { //Home or page_up
 				caret(input, 0, e.shiftKey ? pos.begin : 0);
-			} else if (((opts.undoOnEscape && k == Inputmask.keyCode.ESCAPE) || (k == 90 && e.ctrlKey)) && e.altKey !== true) { //escape && undo && #762
+			} else if (((opts.undoOnEscape && k === Inputmask.keyCode.ESCAPE) || (k === 90 && e.ctrlKey)) && e.altKey !== true) { //escape && undo && #762
 				checkVal(input, true, false, undoValue.split(''));
 				$input.click();
-			} else if (k == Inputmask.keyCode.INSERT && !(e.shiftKey || e.ctrlKey)) { //insert
+			} else if (k === Inputmask.keyCode.INSERT && !(e.shiftKey || e.ctrlKey)) { //insert
 				opts.insertMode = !opts.insertMode;
-				caret(input, !opts.insertMode && pos.begin == getMaskLength() ? pos.begin - 1 : pos.begin);
-			} else if (opts.insertMode == false && !e.shiftKey) {
-				if (k == Inputmask.keyCode.RIGHT) {
+				caret(input, !opts.insertMode && pos.begin === getMaskLength() ? pos.begin - 1 : pos.begin);
+			} else if (opts.tabThrough === true && k === Inputmask.keyCode.TAB) {
+				if (e.shiftKey === true) {
+					if (getTest(pos.begin).fn === null)
+						pos.begin = seekNext(pos.begin);
+					pos.end = seekPrevious(pos.begin, true);
+					pos.begin = seekPrevious(pos.end, true);
+				} else {
+					pos.begin = seekNext(pos.begin, true);
+					pos.end = seekNext(pos.begin, true);
+					if (pos.end < getMaskLength()) pos.end--;
+				}
+				if (pos.begin < getMaskLength()) {
+					e.preventDefault();
+					caret(input, pos.begin, pos.end);
+				}
+			} else if (opts.insertMode === false && !e.shiftKey) {
+				if (k === Inputmask.keyCode.RIGHT) {
 					setTimeout(function() {
 						var caretPos = caret(input);
 						caret(input, caretPos.begin);
 					}, 0);
-				} else if (k == Inputmask.keyCode.LEFT) {
+				} else if (k === Inputmask.keyCode.LEFT) {
 					setTimeout(function() {
 						var caretPos = caret(input);
 						caret(input, isRTL ? caretPos.begin + 1 : caretPos.begin - 1);
@@ -2042,7 +2058,7 @@
 				}
 			}
 			opts.onKeyDown.call(this, e, getBuffer(), caret(input).begin, opts);
-			ignorable = $.inArray(k, opts.ignorables) != -1;
+			ignorable = $.inArray(k, opts.ignorables) !== -1;
 		}
 
 		function keypressEvent(e, checkval, writeOut, strict, ndx) {
@@ -2355,7 +2371,7 @@
 						input = this;
 					if ($input.is(":focus")) {
 						var selectedCaret = caret(input);
-						if (selectedCaret.begin == selectedCaret.end) {
+						if (selectedCaret.begin === selectedCaret.end) {
 							if (opts.radixFocus && opts.radixPoint != "" && $.inArray(opts.radixPoint, getBuffer()) != -1 && (firstClick || getBuffer().join('') == getBufferTemplate().join(''))) {
 								caret(input, $.inArray(opts.radixPoint, getBuffer()));
 								firstClick = false;
@@ -2371,12 +2387,12 @@
 							}
 						}
 					}
-				}).bind('dblclick.inputmask', function() {
+				}).bind("dblclick.inputmask", function() {
 					var input = this;
 					setTimeout(function() {
 						caret(input, 0, seekNext(getLastValidPosition()));
 					}, 0);
-				}).bind(PasteEventType + ".inputmask dragdrop.inputmask drop.inputmask", pasteEvent).bind('cut.inputmask', function(e) {
+				}).bind(PasteEventType + ".inputmask dragdrop.inputmask drop.inputmask", pasteEvent).bind("cut.inputmask", function(e) {
 					skipInputEvent = true; //stop inputFallback
 					var input = this,
 						$input = $(input),
@@ -2385,19 +2401,19 @@
 					//correct clipboardData
 					if (isRTL) {
 						var clipboardData = window.clipboardData || e.originalEvent.clipboardData,
-							clipData = clipboardData.getData('text').split('').reverse().join('');
-						clipboardData.setData('text', clipData);
+							clipData = clipboardData.getData("text").split('').reverse().join('');
+						clipboardData.setData("text", clipData);
 					}
 					handleRemove(input, Inputmask.keyCode.DELETE, pos);
-					writeBuffer(input, getBuffer(), getMaskSet()["p"], e, undoValue != getBuffer().join(''));
+					writeBuffer(input, getBuffer(), getMaskSet().p, e, undoValue != getBuffer().join(''));
 
 					if (input.inputmask._valueGet() == getBufferTemplate().join(''))
-						$input.trigger('cleared');
+						$input.trigger("cleared");
 
 					if (opts.showTooltip) { //update tooltip
-						$input.prop("title", getMaskSet()["mask"]);
+						$input.prop("title", getMaskSet().mask);
 					}
-				}).bind('complete.inputmask', opts.oncomplete).bind('incomplete.inputmask', opts.onincomplete).bind('cleared.inputmask', opts.oncleared);
+				}).bind("complete.inputmask", opts.oncomplete).bind("incomplete.inputmask", opts.onincomplete).bind("cleared.inputmask", opts.oncleared);
 
 				$el.bind("keydown.inputmask", keydownEvent).bind("keypress.inputmask", keypressEvent);
 
@@ -2414,7 +2430,7 @@
 				//}
 			}
 
-			$el.bind('setvalue.inputmask', function() {
+			$el.bind("setvalue.inputmask", function() {
 				var input = this,
 					value = input.inputmask._valueGet();
 				input.inputmask._valueSet($.isFunction(opts.onBeforeMask) ? (opts.onBeforeMask.call(input, value, opts) || value) : value);
@@ -2456,18 +2472,18 @@
 		}
 
 		//action object
-		if (actionObj != undefined) {
+		if (actionObj !== undefined) {
 			switch (actionObj["action"]) {
 				case "isComplete":
-					el = actionObj["el"]
+					el = actionObj.el;
 					$el = $(el);
 					maskset = el.inputmask.maskset;
 					opts = el.inputmask.opts;
-					return isComplete(actionObj["buffer"]);
+					return isComplete(actionObj.buffer);
 				case "unmaskedvalue":
-					el = actionObj["el"]
+					el = actionObj.el;
 
-					if (el == undefined) {
+					if (el === undefined) {
 						//store inputmask instance on the input with element reference
 						$el = $({});
 						el = $el[0];
@@ -2480,7 +2496,7 @@
 						if (opts.numericInput) {
 							isRTL = true;
 						}
-						var valueBuffer = ($.isFunction(opts.onBeforeMask) ? (opts.onBeforeMask.call($el, actionObj["value"], opts) || actionObj["value"]) : actionObj["value"]).split('');
+						var valueBuffer = ($.isFunction(opts.onBeforeMask) ? (opts.onBeforeMask.call($el, actionObj.value, opts) || actionObj.value) : actionObj.value).split('');
 						checkVal($el, false, false, isRTL ? valueBuffer.reverse() : valueBuffer);
 						$.isFunction(opts.onBeforeWrite) && opts.onBeforeWrite.call(this, undefined, getBuffer(), 0, opts);
 					} else $el = $(el);
@@ -2490,7 +2506,7 @@
 					return unmaskedvalue($el);
 				case "mask":
 					undoValue = getBuffer().join('');
-					mask(actionObj["el"]);
+					mask(actionObj.el);
 					break;
 				case "format":
 					$el = $({});
@@ -2504,11 +2520,11 @@
 					if (opts.numericInput) {
 						isRTL = true;
 					}
-					var valueBuffer = ($.isFunction(opts.onBeforeMask) ? (opts.onBeforeMask.call($el, actionObj["value"], opts) || actionObj["value"]) : actionObj["value"]).split('');
+					var valueBuffer = ($.isFunction(opts.onBeforeMask) ? (opts.onBeforeMask.call($el, actionObj.value, opts) || actionObj.value) : actionObj.value).split('');
 					checkVal($el, false, false, isRTL ? valueBuffer.reverse() : valueBuffer);
 					$.isFunction(opts.onBeforeWrite) && opts.onBeforeWrite.call(this, undefined, getBuffer(), 0, opts);
 
-					if (actionObj["metadata"]) {
+					if (actionObj.metadata) {
 						return {
 							value: isRTL ? getBuffer().slice().reverse().join('') : getBuffer().join(''),
 							metadata: $el.inputmask("getmetadata")
@@ -2528,7 +2544,7 @@
 					if (opts.numericInput) {
 						isRTL = true;
 					}
-					var valueBuffer = actionObj["value"].split('');
+					var valueBuffer = actionObj.value.split('');
 					checkVal($el, false, true, isRTL ? valueBuffer.reverse() : valueBuffer);
 					var buffer = getBuffer();
 					var rl = determineLastRequiredPosition(),
@@ -2538,9 +2554,9 @@
 					}
 					buffer.splice(rl, lmib + 1 - rl);
 
-					return isComplete(buffer) && actionObj["value"] == buffer.join('');
+					return isComplete(buffer) && actionObj.value == buffer.join('');
 				case "getemptymask":
-					el = actionObj["el"]
+					el = actionObj.el;
 					$el = $(el);
 					maskset = el.inputmask.maskset;
 					opts = el.inputmask.opts;